test_requirements.txt
coverage
fabric3
flake8
ipdb
mixer
mock
tox
django-compat
six
